pyproject.toml
[project]
name = "leopoliscampfeedback"
version = "0.1.0"
description = ""
authors = [
    {name = "Oleh Andrus",email = "andrusoleg123@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-auth (>=2.40.3,<3.0.0)",
    "gspread (>=6.2.1,<7.0.0)",
    "python-telegram-bot (>=22.1,<23.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "python-i18n[yaml] (>=0.3.9,<0.4.0)",
    "pre-commit (>=4.2.0,<5.0.0)",
    "black (>=25.1.0,<26.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
